fix auth in api routes

+++ routes/api.php
@@ -14,15 +14,16 @@ use Illuminate\Http\Request;
 */
 
 Route::post('register',  'Api\AuthController@register');
-Route::middleware('api')->prefix('movie')->group(function () {
-    Route::get('/', 'Api\TMDBController@search');
-    Route::get('/{id}', 'Api\TMDBController@show')->where('id', '[0-9]+');
-    Route::get('/popular', 'Api\TMDBController@popular');
+
+Route::prefix('movie')->group(function () {
+    Route::get('/{id}', 'Api\TMDBController@show')->where('id', '[0-9]+')->middleware('auth:api');
+    Route::get('/search', 'Api\TMDBController@search')->middleware('auth:api');
+    Route::get('/popular', 'Api\TMDBController@popular')->middleware('auth:api');
 });
 
-Route::middleware('api')->prefix('user')->group(function () {
-    Route::get('/favorites', 'Api\UserController@favorites');
-    Route::post('/favorites/{id}', 'Api\UserController@addFavourite')->where('id', '[0-9]+');
-    Route::delete('/favorites/{id}', 'Api\UserController@removeFavourite')->where('id', '[0-9]+');
+Route::prefix('user')->group(function () {
+    Route::get('/favorites', 'Api\UserController@favorites')->middleware('auth:api');
+    Route::post('/favorites/{id}', 'Api\UserController@addFavourite')->where('id', '[0-9]+')->middleware('auth:api');
+    Route::delete('/favorites/{id}', 'Api\UserController@removeFavourite')->where('id', '[0-9]+')->middleware('auth:api');
 });
